+++ ocrd_segment/cli.py
@@ -8,6 +8,7 @@ from .import_coco_segmentation import ImportCOCOSegmentation
 from .evaluate import EvaluateSegmentation
 from .replace_original import ReplaceOriginal
 from .replace_page import ReplacePage
+from .replace_text import ReplaceText
 from .extract_pages import ExtractPages
 from .extract_regions import ExtractRegions
 from .extract_lines import ExtractLines
@@ -56,6 +57,11 @@ def ocrd_segment_replace_original(*args, **kwargs):
 @ocrd_cli_options
 def ocrd_segment_replace_page(*args, **kwargs):
     return ocrd_cli_wrap_processor(ReplacePage, *args, **kwargs)
+
+@click.command()
+@ocrd_cli_options
+def ocrd_segment_replace_text(*args, **kwargs):
+    return ocrd_cli_wrap_processor(ReplaceText, *args, **kwargs)
 
 @click.command()
 @ocrd_cli_options